Service_System/pyproject.toml
[project]
name = "photoblogserver"
version = "0.1.0"
description = "Photo Blog Server with Smart Bollard System"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.0.0",
    "pillow>=12.0.0",
    "grpcio>=1.60.0",
    "requests>=2.31.0",
]
